added individual invoice page to view each invoice, added logic to dynamically show data and fixed some number calculating bugs

+++ src/routes/(dashboard)/invoices/LineItemRow.svelte
@@ -1,14 +1,19 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
 	import Trash from '$lib/components/Icon/Trash.svelte';
-	import { twoDecimals, dollarsToCents, centsToDollars } from '$lib/utils/moneyHelpers';
+	import {
+		twoDecimals,
+		dollarsToCents,
+		centsToDollarsWithoutCommas
+	} from '$lib/utils/moneyHelpers';
 
 	export let lineItem: LineItem;
 	export let canDelete: boolean = false;
 	export let isRequired: boolean = false;
+	export let isEditable: boolean = true;
 
-	let unitPrice: string = centsToDollars(lineItem.amount / lineItem.quantity);
-	let amount: string = centsToDollars(lineItem.amount);
+	let unitPrice: string = centsToDollarsWithoutCommas(lineItem.amount / lineItem.quantity);
+	let amount: string = centsToDollarsWithoutCommas(lineItem.amount);
 
 	$: {
 		amount = twoDecimals(lineItem.quantity * Number(unitPrice));
@@ -27,6 +32,7 @@
 			name="description"
 			bind:value={lineItem.description}
 			required={isRequired}
+			disabled={!isEditable}
 		/>
 	</div>
 
@@ -38,6 +44,7 @@
 			name="unitPrice"
 			step="0.01"
 			min="0"
+			disabled={!isEditable}
 			bind:value={unitPrice}
 			on:blur={() => {
 				unitPrice = twoDecimals(Number(unitPrice));
@@ -54,6 +61,7 @@
 			type="number"
 			name="quantity"
 			min="0"
+			disabled={!isEditable}
 			bind:value={lineItem.quantity}
 			on:blur={() => {
 				dispatch('updateLineItem');
@@ -76,7 +84,7 @@
 	</div>
 
 	<div class="trash">
-		{#if canDelete}
+		{#if canDelete && isEditable}
 			<button
 				class="center h-10 w-10 text-pastelPurple hover:text-lavenderIndigo"
 				on:click|preventDefault={() => dispatch('removeLineItem', lineItem.id)}
